requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.17.1
python-telegram-bot==13.6
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
